Crate ltsinfo_lts

A crate containing labelled transition systems related functionality.

This crate does not use unsafe code.

Structs

CompactTransition

A compact representation of a transition using a single u64. The high 16 bits store the label index, the low 48 bits store the target state index.

LabelledTransitionSystem

Represents a labelled transition system consisting of states with directed labelled edges.

Functions

random_lts

Generates a monolithic LTS with the desired number of states, labels, out degree and in degree for all the states.

Type Aliases

LabelIndex

The index type for a label.

StateIndex

The index for a state.